renamed package pixelgl to opengl
