+++ backends/opengl/canvas.go
@@ -1,4 +1,4 @@
-package pixelgl
+package opengl
 
 import (
 	"fmt"
