+++ backends/opengl/monitor.go
@@ -1,4 +1,4 @@
-package pixelgl
+package opengl
 
 import (
 	"github.com/go-gl/glfw/v3.3/glfw"
